console: Adjust prop-types for events subscription
